Initial add of refactored dd_send
